refactor: dispatch chat callbacks through ccstate
