+++ turbo/apps/api/src/signals/routes/__tests__/chat-callbacks.bdd.test.ts
@@ -7,6 +7,7 @@ import type {
   GenerationTemplateRequest,
   PagedChatMessage,
 } from "@vm0/api-contracts/contracts/chat-threads";
+import { agentRunCallbacks } from "@vm0/db/schema/agent-run-callback";
 import { orgMembersCache } from "@vm0/db/schema/org-members-cache";
 import { zeroRuns } from "@vm0/db/schema/zero-run";
 import { createStore } from "ccstate";
@@ -16,6 +17,7 @@ import { describe, expect, it } from "vitest";
 import { mockOptionalEnv } from "../../../lib/env";
 import { nowDate } from "../../../lib/time";
 import { testContext } from "../../../__tests__/test-helpers";
+import { flushWaitUntilForTest } from "../../context/wait-until";
 import { writeDb$ } from "../../external/db";
 import { MODEL_FIRST_SELECTION_PROVIDER_ID } from "../../services/zero-model-selection.service";
 import { createBddApi, type ApiTestUser } from "./helpers/api-bdd";
@@ -23,15 +25,15 @@ import { createChatCallbacksApi } from "./helpers/api-bdd-chat-callbacks";
 import { createChatFilesBddApi } from "./helpers/api-bdd-chat-files";
 import { createRunsAutomationsApi } from "./helpers/api-bdd-runs-automations";
 import { createWebhookCallbackApi } from "./helpers/api-bdd-webhooks";
+import { encryptSecretForTests } from "./helpers/encrypt-secret";
 
 /**
  * CHAT-02 / HOOK-01: signed chat run callbacks through real dispatch.
  *
- * Every signed callback in this file originates from the real dispatcher
- * (sandbox complete/cancel webhooks and the sandbox heartbeat route) and is
- * either proxied into the app or captured for verbatim replay. Tests create
- * runs through public APIs; scenarios about persisted run context may adjust
- * that context directly before sending the real callback.
+ * Terminal callbacks in this file originate from the real internal dispatcher
+ * (sandbox complete/cancel webhooks and the sandbox heartbeat route). The
+ * legacy HTTP wrapper coverage constructs signed callback envelopes explicitly
+ * so normal app-internal dispatch does not depend on an HTTP self-call.
  */
 
 const context = testContext();
@@ -197,17 +199,6 @@ async function seedOrgMemberRole(
     });
 }
 
-async function waitForArrayLength<T>(
-  items: readonly T[],
-  length: number,
-): Promise<void> {
-  await expect
-    .poll(() => {
-      return items.length;
-    })
-    .toBe(length);
-}
-
 async function waitForThreadMessages(
   actor: ApiTestUser,
   threadId: string,
@@ -249,6 +240,45 @@ async function waitForRunStatus(
       return run.status;
     })
     .toBe(status);
+}
+
+async function signedLegacyChatCallbackDelivery(args: {
+  readonly runId: string;
+  readonly threadId: string;
+  readonly agentId: string;
+  readonly status: "completed" | "failed";
+  readonly error?: string;
+}) {
+  const secret = `legacy-chat-callback-${args.runId}`;
+  const db = store.set(writeDb$);
+  const [callback] = await db
+    .select({ id: agentRunCallbacks.id })
+    .from(agentRunCallbacks)
+    .where(eq(agentRunCallbacks.runId, args.runId))
+    .limit(1);
+  if (!callback) {
+    throw new Error("Expected a chat callback row");
+  }
+
+  await db
+    .update(agentRunCallbacks)
+    .set({
+      url: "http://localhost:3000/api/internal/callbacks/chat",
+      internalKind: null,
+      encryptedSecret: encryptSecretForTests(secret),
+    })
+    .where(eq(agentRunCallbacks.id, callback.id));
+
+  return chatCallbacks.signedChatCallbackDelivery(
+    {
+      callbackId: callback.id,
+      runId: args.runId,
+      status: args.status,
+      ...(args.error === undefined ? {} : { error: args.error }),
+      payload: { threadId: args.threadId, agentId: args.agentId },
+    },
+    secret,
+  );
 }
 
 /**
@@ -362,7 +392,7 @@ function pushPayload(call: readonly unknown[] | undefined): unknown {
 describe("CHAT-02: completed chat callback", () => {
   it("persists assistant output, reorders threads, titles the thread, recommends follow-ups, notifies, and auto-sends the queued template message", async () => {
     const { actor, agentId, runnerGroup } = await entitledChatActor();
-    chatCallbacks.proxyChatCallbackToApp();
+    chatCallbacks.failIfChatCallbackRouteIsFetched();
 
     const titlePrompts: string[] = [];
     mockOptionalEnv("OPENROUTER_API_KEY", "bdd-openrouter-key");
@@ -617,7 +647,7 @@ describe("CHAT-02: completed chat callback", () => {
 describe("CHAT-02: chat output extraction and progress callbacks", () => {
   it("extracts assistant output from Codex items and result fallbacks, skips non-events, and acknowledges progress without reading events", async () => {
     const { actor, agentId, runnerGroup } = await entitledChatActor();
-    const deliveries = chatCallbacks.proxyChatCallbackToApp();
+    const routeRequests = chatCallbacks.failIfChatCallbackRouteIsFetched();
 
     const first = await startChatRun(actor, {
       agentId,
@@ -652,14 +682,7 @@ describe("CHAT-02: chat output extraction and progress callbacks", () => {
       [200],
     );
 
-    await waitForArrayLength(deliveries, 1);
-    const progressBody: unknown = JSON.parse(deliveries[0]?.body ?? "{}");
-    expect(progressBody).toMatchObject({
-      callbackId: expect.stringMatching(/[0-9a-f-]{36}/),
-      runId: first.runId,
-      status: "progress",
-      payload: { threadId: first.threadId, agentId },
-    });
+    expect(routeRequests()).toBe(0);
     expect(context.mocks.axiom.query).not.toHaveBeenCalled();
     expect(context.mocks.ably.publish).not.toHaveBeenCalledWith(
       `chatThreadMessageCreated:${first.threadId}`,
@@ -816,7 +839,6 @@ describe("CHAT-02: chat output extraction and progress callbacks", () => {
 describe("CHAT-02/HOOK-01: chat callback replay and signature handling", () => {
   it("deduplicates concurrent and replayed deliveries and rejects tampered signatures", async () => {
     const { actor, agentId, runnerGroup } = await entitledChatActor();
-    const deliveries = chatCallbacks.captureChatCallbackDeliveries();
 
     const first = await startChatRun(actor, { agentId, prompt: "dedupe me" });
     const firstHeaders = await claimChatRun(runnerGroup, first.runId);
@@ -828,30 +850,6 @@ describe("CHAT-02/HOOK-01: chat callback replay and signature handling", () => {
       lastEventSequence: 1,
     });
 
-    await waitForArrayLength(deliveries, 1);
-    const completedDelivery = deliveries[0];
-    if (!completedDelivery) {
-      throw new Error("Expected a captured completed delivery");
-    }
-    const completedBody: unknown = JSON.parse(completedDelivery.body);
-    expect(completedBody).toMatchObject({
-      callbackId: expect.stringMatching(/[0-9a-f-]{36}/),
-      runId: first.runId,
-      status: "completed",
-      payload: { threadId: first.threadId, agentId },
-    });
-    expect(completedDelivery.headers["x-vm0-signature"]).toMatch(/.+/);
-    expect(completedDelivery.headers["x-vm0-timestamp"]).toMatch(/^\d+$/);
-
-    const beforeReplay = await chat.listThreadMessages(actor, first.threadId);
-    expect(assistantMessages(beforeReplay.messages)).toHaveLength(0);
-
-    const [replayA, replayB] = await Promise.all([
-      chatCallbacks.replayChatCallback(completedDelivery),
-      chatCallbacks.replayChatCallback(completedDelivery),
-    ]);
-    expect(replayA.status).toBe(200);
-    expect(replayB.status).toBe(200);
     let messages = await waitForThreadMessages(
       actor,
       first.threadId,
@@ -874,14 +872,63 @@ describe("CHAT-02/HOOK-01: chat callback replay and signature handling", () => {
       lifecycleMarkers(messages.messages, first.runId, "completed"),
     ).toHaveLength(1);
 
+    const completedDelivery = await signedLegacyChatCallbackDelivery({
+      runId: first.runId,
+      threadId: first.threadId,
+      agentId,
+      status: "completed",
+    });
+    const completedBody: unknown = JSON.parse(completedDelivery.body);
+    expect(completedBody).toMatchObject({
+      callbackId: expect.stringMatching(/[0-9a-f-]{36}/),
+      runId: first.runId,
+      status: "completed",
+      payload: { threadId: first.threadId, agentId },
+    });
+    expect(completedDelivery.headers["x-vm0-signature"]).toMatch(/.+/);
+    expect(completedDelivery.headers["x-vm0-timestamp"]).toMatch(/^\d+$/);
+
     const sentinel = await chat.createThread(actor, {
       agentId,
       title: "Replay ordering sentinel",
     });
     context.mocks.ably.publish.mockClear();
+    const [replayA, replayB] = await Promise.all([
+      chatCallbacks.replayChatCallback(completedDelivery),
+      chatCallbacks.replayChatCallback(completedDelivery),
+    ]);
+    expect(replayA.status).toBe(200);
+    expect(replayB.status).toBe(200);
+    await flushWaitUntilForTest();
+    messages = await waitForThreadMessages(
+      actor,
+      first.threadId,
+      (threadMessages) => {
+        return (
+          eventBackedContents(threadMessages, first.runId).length === 2 &&
+          lifecycleMarkers(threadMessages, first.runId, "completed").length ===
+            1
+        );
+      },
+    );
+    expect(
+      eventBackedContents(messages.messages, first.runId)
+        .map((message) => {
+          return message.content;
+        })
+        .sort(),
+    ).toStrictEqual(["First event", "Second event"]);
+    expect(
+      lifecycleMarkers(messages.messages, first.runId, "completed"),
+    ).toHaveLength(1);
+    expect(context.mocks.ably.publish).not.toHaveBeenCalledWith(
+      `chatThreadMessageCreated:${first.threadId}`,
+      null,
+    );
     const replayAgain =
       await chatCallbacks.replayChatCallback(completedDelivery);
     expect(replayAgain.status).toBe(200);
+    await flushWaitUntilForTest();
     messages = await waitForThreadMessages(
       actor,
       first.threadId,
@@ -933,11 +980,18 @@ describe("CHAT-02/HOOK-01: chat callback replay and signature handling", () => {
       "Cannot continue session from checkpoint",
     );
 
-    await waitForArrayLength(deliveries, 2);
-    const failedDelivery = deliveries[1];
-    if (!failedDelivery) {
-      throw new Error("Expected a captured failed delivery");
-    }
+    await waitForThreadMessages(actor, first.threadId, (threadMessages) => {
+      return (
+        lifecycleMarkers(threadMessages, second.runId, "failed").length === 1
+      );
+    });
+    const failedDelivery = await signedLegacyChatCallbackDelivery({
+      runId: second.runId,
+      threadId: first.threadId,
+      agentId,
+      status: "failed",
+      error: "Cannot continue session from checkpoint",
+    });
     const failedBody: unknown = JSON.parse(failedDelivery.body);
     expect(failedBody).toMatchObject({
       runId: second.runId,
@@ -946,18 +1000,15 @@ describe("CHAT-02/HOOK-01: chat callback replay and signature handling", () => {
       payload: { threadId: first.threadId, agentId },
     });
 
+    context.mocks.ably.publish.mockClear();
     const firstFailedReplay =
       await chatCallbacks.replayChatCallback(failedDelivery);
     expect(firstFailedReplay.status).toBe(200);
-    await waitForThreadMessages(actor, first.threadId, (threadMessages) => {
-      return (
-        lifecycleMarkers(threadMessages, second.runId, "failed").length === 1
-      );
-    });
-    context.mocks.ably.publish.mockClear();
+    await flushWaitUntilForTest();
     const secondFailedReplay =
       await chatCallbacks.replayChatCallback(failedDelivery);
     expect(secondFailedReplay.status).toBe(200);
+    await flushWaitUntilForTest();
     messages = await waitForThreadMessages(
       actor,
       first.threadId,
@@ -986,7 +1037,7 @@ describe("CHAT-02/HOOK-01: chat callback replay and signature handling", () => {
 describe("CHAT-02: failed chat callbacks", () => {
   it("formats failed-run errors with escalation and notifies, without auto-sending", async () => {
     const { actor, agentId, runnerGroup } = await entitledChatActor();
-    chatCallbacks.proxyChatCallbackToApp();
+    chatCallbacks.failIfChatCallbackRouteIsFetched();
     await chatCallbacks.registerPushSubscription(actor);
     chatCallbacks.enableVapid();
 
@@ -1087,7 +1138,7 @@ describe("CHAT-02: failed chat callbacks", () => {
 
   it("shows Claude Code credential recovery guidance for upstream auth 401s", async () => {
     const { actor, agentId, runnerGroup } = await entitledChatActor();
-    chatCallbacks.proxyChatCallbackToApp();
+    chatCallbacks.failIfChatCallbackRouteIsFetched();
     const upstreamAuthError =
       "Failed to authenticate. API Error: 401 Invalid authentication credentials";
 
@@ -1166,7 +1217,7 @@ describe("CHAT-02: failed chat callbacks", () => {
 describe("CHAT-02: auto-send after failures", () => {
   it("auto-sends the queued message after a failure, carrying attachments, incomplete-round context, and the continued session", async () => {
     const { actor, agentId, runnerGroup, storage } = await entitledChatActor();
-    chatCallbacks.proxyChatCallbackToApp();
+    chatCallbacks.failIfChatCallbackRouteIsFetched();
 
     const first = await startChatRun(actor, {
       agentId,
@@ -1303,7 +1354,7 @@ describe("CHAT-02: auto-send across a model switch", () => {
   it("starts a fresh session with prior web context when the queued model differs, without regenerating an existing title", async () => {
     const { actor, agentId, runnerGroup, providerId } =
       await entitledChatActor();
-    chatCallbacks.proxyChatCallbackToApp();
+    chatCallbacks.failIfChatCallbackRouteIsFetched();
     await chatCallbacks.updateOrgModelPolicies(actor, [
       {
         model: "claude-sonnet-4-6",
@@ -1437,7 +1488,7 @@ describe("CHAT-02: auto-send across a model switch", () => {
 describe("CHAT-02: thread deletion while a run is active", () => {
   it("skips terminal processing when the thread is gone", async () => {
     const { actor, agentId, runnerGroup } = await entitledChatActor();
-    chatCallbacks.proxyChatCallbackToApp();
+    chatCallbacks.failIfChatCallbackRouteIsFetched();
 
     const run = await startChatRun(actor, {
       agentId,
@@ -1480,7 +1531,7 @@ describe("CHAT-02: thread deletion while a run is active", () => {
 describe("CHAT-02: push notification gating", () => {
   it("withholds pushes without VAPID keys and deletes stale subscriptions after gone responses", async () => {
     const { actor, agentId, runnerGroup } = await entitledChatActor();
-    chatCallbacks.proxyChatCallbackToApp();
+    chatCallbacks.failIfChatCallbackRouteIsFetched();
     const endpoint = await chatCallbacks.registerPushSubscription(actor);
 
     const first = await startChatRun(actor, {
