+++ turbo/apps/api/src/signals/routes/__tests__/helpers/api-bdd-github.ts
@@ -54,7 +54,6 @@ const GITHUB_APP_ID = "123456";
 const DEFAULT_TEST_ORIGIN = "http://localhost:3000";
 const GITHUB_ISSUES_CALLBACK_PATH = "/api/internal/callbacks/github/issues";
 const GITHUB_ISSUES_CALLBACK_URL = `${DEFAULT_TEST_ORIGIN}${GITHUB_ISSUES_CALLBACK_PATH}`;
-const CHAT_CALLBACK_URL = `${DEFAULT_TEST_ORIGIN}/api/internal/callbacks/chat`;
 
 type ComposeContent = z.infer<typeof agentComposeApiContentSchema>;
 
@@ -107,12 +106,6 @@ interface GithubIssueHistoryComment {
   readonly id: number;
   readonly login: string;
   readonly body: string;
-}
-
-interface RecordedCallbackDelivery {
-  readonly body: string;
-  readonly signature: string | null;
-  readonly timestamp: string | null;
 }
 
 interface SignedConnectLink {
@@ -509,26 +502,6 @@ export function proxyGithubIssuesCallbackToApp(context: TestContext): void {
       });
     }),
   );
-}
-
-/**
- * Record signed chat-callback deliveries without proxying them, so a chat
- * run's delivery can be replayed against the GitHub issues callback route
- * (per-callback signature verifies, payload schema does not).
- */
-export function captureChatCallbackDeliveries(): RecordedCallbackDelivery[] {
-  const deliveries: RecordedCallbackDelivery[] = [];
-  server.use(
-    http.post(CHAT_CALLBACK_URL, async ({ request }) => {
-      deliveries.push({
-        body: await request.text(),
-        signature: request.headers.get("x-vm0-signature"),
-        timestamp: request.headers.get("x-vm0-timestamp"),
-      });
-      return HttpResponse.json({ ok: true });
-    }),
-  );
-  return deliveries;
 }
 
 export function buildLegacySignedState(args: {
