+++ turbo/apps/api/src/signals/routes/__tests__/helpers/api-bdd-webhooks.ts
@@ -87,28 +87,6 @@ interface CapturedInternalCallbackDelivery {
   readonly headers: Record<string, string>;
 }
 
-/**
- * First captured delivery whose JSON envelope carries the given callback
- * status. Throws when no such delivery was dispatched yet.
- */
-export function callbackDeliveryWithStatus(
-  deliveries: readonly CapturedInternalCallbackDelivery[],
-  status: "completed" | "failed" | "progress",
-): CapturedInternalCallbackDelivery {
-  const delivery = deliveries.find((entry) => {
-    const parsed: unknown = JSON.parse(entry.body);
-    return (
-      typeof parsed === "object" &&
-      parsed !== null &&
-      (parsed as { readonly status?: unknown }).status === status
-    );
-  });
-  if (!delivery) {
-    throw new Error(`Expected a captured ${status} callback delivery`);
-  }
-  return delivery;
-}
-
 interface SvixHeaders {
   readonly "svix-id": string;
   readonly "svix-timestamp": string;
