+++ turbo/apps/api/src/signals/services/internal-run-callback.ts
@@ -3,6 +3,7 @@ import { safeUrlParse } from "../utils";
 export const internalRunCallbackKinds = [
   "agent",
   "agentphone",
+  "chat",
   "github:issues",
   "slack:org",
   "telegram",
@@ -38,6 +39,7 @@ function isInternalRunCallbackKind(
   switch (value) {
     case "agent":
     case "agentphone":
+    case "chat":
     case "github:issues":
     case "slack:org":
     case "telegram":
@@ -74,6 +76,9 @@ function legacyInternalRunCallbackKind(
     }
     case "/api/internal/callbacks/agentphone": {
       return "agentphone";
+    }
+    case "/api/internal/callbacks/chat": {
+      return "chat";
     }
     case "/api/internal/callbacks/github/issues": {
       return "github:issues";
